Moved messageParser to more reasonable place
